+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/event/EventDao.java
@@ -1,9 +1,9 @@
-package com.walmartlabs.concord.server.project.event;
+package com.walmartlabs.concord.server.process.event;
 
 import com.walmartlabs.concord.common.db.AbstractDao;
 import com.walmartlabs.concord.server.api.process.ProcessEventEntry;
 import com.walmartlabs.concord.server.api.process.ProcessEventType;
-import com.walmartlabs.concord.server.jooq.public_.tables.records.ProcessEventRecord;
+import com.walmartlabs.concord.server.jooq.public_.tables.records.ProcessEventsRecord;
 import org.jooq.Configuration;
 import org.jooq.DSLContext;
 import org.jooq.impl.DSL;
@@ -14,7 +14,7 @@ import java.sql.Timestamp;
 import java.util.Date;
 import java.util.List;
 
-import static com.walmartlabs.concord.server.jooq.public_.tables.ProcessEvent.PROCESS_EVENT;
+import static com.walmartlabs.concord.server.jooq.public_.Tables.PROCESS_EVENTS;
 
 @Named
 public class EventDao extends AbstractDao {
@@ -27,9 +27,9 @@ public class EventDao extends AbstractDao {
     public List<ProcessEventEntry> list(String instanceId) {
         try (DSLContext tx = DSL.using(cfg)) {
 
-            return tx.selectFrom(PROCESS_EVENT)
-                    .where(PROCESS_EVENT.INSTANCE_ID.eq(instanceId))
-                    .orderBy(PROCESS_EVENT.EVENT_DATE)
+            return tx.selectFrom(PROCESS_EVENTS)
+                    .where(PROCESS_EVENTS.INSTANCE_ID.eq(instanceId))
+                    .orderBy(PROCESS_EVENTS.EVENT_DATE)
                     .fetch(EventDao::toEntry);
         }
     }
@@ -39,14 +39,14 @@ public class EventDao extends AbstractDao {
     }
 
     public void insert(DSLContext tx, String instanceId, ProcessEventType eventType, Date eventDate, String eventData) {
-        tx.insertInto(PROCESS_EVENT)
-                .columns(PROCESS_EVENT.INSTANCE_ID, PROCESS_EVENT.EVENT_TYPE,
-                        PROCESS_EVENT.EVENT_DATE, PROCESS_EVENT.EVENT_DATA)
+        tx.insertInto(PROCESS_EVENTS)
+                .columns(PROCESS_EVENTS.INSTANCE_ID, PROCESS_EVENTS.EVENT_TYPE,
+                        PROCESS_EVENTS.EVENT_DATE, PROCESS_EVENTS.EVENT_DATA)
                 .values(instanceId, eventType.name(), new Timestamp(eventDate.getTime()), eventData)
                 .execute();
     }
 
-    private static ProcessEventEntry toEntry(ProcessEventRecord r) {
+    private static ProcessEventEntry toEntry(ProcessEventsRecord r) {
         return new ProcessEventEntry(ProcessEventType.valueOf(r.getEventType()), r.getEventDate(), r.getEventData());
     }
 }